+++ carpetmodSrc/carpet/utils/Messenger.java
@@ -358,11 +358,20 @@ public class Messenger {
      * Using "/tp" instead of "!/tp"
      */
     @Nonnull
-    public static ITextComponent tpa(String style, double x, double y, double z) {
-        String display = String.format("%s [%.1f, %.1f, %.1f]", style, x, y, z);
-        String detail = String.format("^g " + x + ", " + y + ", " + z);
-        String suggest = String.format("/tp " + x + ' ' + y + ' ' + z);
+    public static ITextComponent tpa(String style, double x, double y, double z, int decimals) {
+        String formatter = String.format("%%.%df", decimals);
+        String display = String.format("%s [" + formatter + ", " + formatter + ", " + formatter + "]", style, x, y, z);
+        String detail = "^g " + x + ", " + y + ", " + z;
+        String suggest = "/tp " + x + ' ' + y + ' ' + z;
         return m(null, display, detail, suggest);
+    }
+
+    /**
+     * Using "/tp" instead of "!/tp"
+     */
+    @Nonnull
+    public static ITextComponent tpa(String style, double x, double y, double z) {
+        return tpa(style, x, y, z, 1);
     }
 
     /**
@@ -371,7 +380,7 @@ public class Messenger {
     @Nonnull
     public static ITextComponent tpa(String style, int x, int y, int z) {
         String display = String.format("%s [%d, %d, %d]", style, x, y, z);
-        String suggest = String.format("/tp " + x + ' ' + y + ' ' + z);
-        return m(null, display, suggest);
+        String suggest = "/tp " + x + ' ' + y + ' ' + z;
+        return m(null, display, "^g click to teleport", suggest);
     }
 }